Restructure packaging for sepa
COAND-497
