+++ sepa/src/main/java/com/adyen/checkout/sepa/internal/ui/SepaDelegate.kt
@@ -1,4 +1,12 @@
-package com.adyen.checkout.sepa
+/*
+ * Copyright (c) 2022 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ *
+ * Created by ozgur on 19/7/2022.
+ */
+
+package com.adyen.checkout.sepa.internal.ui
 
 import com.adyen.checkout.components.PaymentComponentState
 import com.adyen.checkout.components.base.PaymentComponentDelegate
@@ -6,9 +14,11 @@ import com.adyen.checkout.components.model.payments.request.SepaPaymentMethod
 import com.adyen.checkout.components.ui.ButtonDelegate
 import com.adyen.checkout.components.ui.UIStateDelegate
 import com.adyen.checkout.components.ui.ViewProvidingDelegate
+import com.adyen.checkout.sepa.internal.ui.model.SepaInputData
+import com.adyen.checkout.sepa.internal.ui.model.SepaOutputData
 import kotlinx.coroutines.flow.Flow
 
-interface SepaDelegate :
+internal interface SepaDelegate :
     PaymentComponentDelegate<PaymentComponentState<SepaPaymentMethod>>,
     ViewProvidingDelegate,
     ButtonDelegate,
