+++ sepa/src/main/java/com/adyen/checkout/sepa/internal/ui/view/IbanInput.kt
@@ -5,16 +5,19 @@
  *
  * Created by caiof on 28/8/2019.
  */
-package com.adyen.checkout.sepa.ui
+package com.adyen.checkout.sepa.internal.ui.view
 
 import android.content.Context
 import android.text.Editable
 import android.util.AttributeSet
 import com.adyen.checkout.components.ui.view.AdyenTextInputEditText
-import com.adyen.checkout.sepa.Iban
+import com.adyen.checkout.sepa.internal.ui.model.Iban
 
-class IbanInput @JvmOverloads constructor(context: Context, attrs: AttributeSet? = null, defStyleAttr: Int = 0) :
-    AdyenTextInputEditText(context, attrs, defStyleAttr) {
+internal class IbanInput @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0
+) : AdyenTextInputEditText(context, attrs, defStyleAttr) {
 
     override fun afterTextChanged(editable: Editable) {
         val initial = editable.toString()
