+++ sepa/src/test/java/com/adyen/checkout/sepa/internal/ui/DefaultSepaDelegateTest.kt
@@ -1,4 +1,12 @@
-package com.adyen.checkout.sepa
+/*
+ * Copyright (c) 2022 Adyen N.V.
+ *
+ * This file is open source and available under the MIT license. See the LICENSE file for more info.
+ *
+ * Created by ozgur on 21/7/2022.
+ */
+
+package com.adyen.checkout.sepa.internal.ui
 
 import app.cash.turbine.test
 import com.adyen.checkout.components.PaymentComponentState
@@ -11,6 +19,8 @@ import com.adyen.checkout.components.model.payments.request.SepaPaymentMethod
 import com.adyen.checkout.components.repository.PaymentObserverRepository
 import com.adyen.checkout.components.ui.SubmitHandler
 import com.adyen.checkout.core.api.Environment
+import com.adyen.checkout.sepa.SepaConfiguration
+import com.adyen.checkout.sepa.internal.ui.model.SepaOutputData
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
